* wrappers/w3af/ospd-w3af.py (OSPDw3af [exec_scan]): Don't remove output file in debugging mode.
git-svn-id: svn+ssh://scm.wald.intevation.org/openvas/trunk/ospd@20394 423fd1db-d629-0410-8442-d21db03e70f4

+++ wrappers/w3af/ospd-w3af.py
@@ -205,7 +205,10 @@ class OSPDw3af(OSPDaemon):
         self.store_scan_results(scan_id, output_file)
 
         # Cleanup
-        os.remove(output_file)
+        if self.logger.get_level() < 1:
+            os.remove(output_file)
+        else:
+            self.logger.debug(2, "{0} not removed.".format(output_file))
         os.remove(script_file)
         # Set scan as finished
         self.finish_scan(scan_id)
